+++ sdks/typescript/client/src/types.ts
@@ -1,12 +1,16 @@
 import { RemoteReceiver } from '@remote-dom/core/receivers';
 import React from 'react';
 
-export type UIActionType = 'tool' | 'prompt' | 'link' | 'intent' | 'notification';
+export type UIActionType = 'tool' | 'prompt' | 'link' | 'intent' | 'notify';
 
 export const ALL_RESOURCE_CONTENT_TYPES = ['rawHtml', 'externalUrl', 'remoteDom'] as const;
 export type ResourceContentType = (typeof ALL_RESOURCE_CONTENT_TYPES)[number];
 
-export type UIActionResultToolCall = {
+type GenericActionMessage = {
+  messageId?: string;
+};
+
+export type UIActionResultToolCall = GenericActionMessage & {
   type: 'tool';
   payload: {
     toolName: string;
@@ -14,21 +18,21 @@ export type UIActionResultToolCall = {
   };
 };
 
-export type UIActionResultPrompt = {
+export type UIActionResultPrompt = GenericActionMessage & {
   type: 'prompt';
   payload: {
     prompt: string;
   };
 };
 
-export type UIActionResultLink = {
+export type UIActionResultLink = GenericActionMessage & {
   type: 'link';
   payload: {
     url: string;
   };
 };
 
-export type UIActionResultIntent = {
+export type UIActionResultIntent = GenericActionMessage & {
   type: 'intent';
   payload: {
     intent: string;
@@ -36,8 +40,8 @@ export type UIActionResultIntent = {
   };
 };
 
-export type UIActionResultNotification = {
-  type: 'notification';
+export type UIActionResultNotification = GenericActionMessage & {
+  type: 'notify';
   payload: {
     message: string;
   };